Branch Logic
We also need to know the target PC of the branch instruction. The target PC is given in the immediate field as a relative byte offset from the current PC. So, the target PC is the PC of the branch plus its immediate value.

Code an expression for $br_tgt_pc[31:0] = …

If the instruction is a taken branch, its next PC should be the branch target PC. Update the existing $next_pc expression to reflect this.

Compile/simulate and debug. Once all is correct, your program will be looping. It should stop looping once it has produced the sum of values 1..9 (45). The final ADDI subtracts 44 from this and should therefore produce a value of 1 in x30. Then the final BGE should loop indefinitely to itself.