ALU Complete, cycle from 50 to 70, error at x24(gives 81 instead of 1)
create assignments for these subexpressions
Extend the expression for $result to complete the ALU to support the remaining instructions.

Compile/simulate, resolve any unexpected issues in the LOG. $is_<instr> signals should no longer be unused.

If any of these new instructions are not resulting in register values of 1 in VIZ, debug them. To be specific, at the end of the simulation, register values should be 1 except x0-4, x27, and x31. Save your work outside of Makerchip.

cycle from 50 to 70

still error at x24(gives 81 instead of 1)