Jump Logic
The ISA, in addition to conditional branches, also supports jump instructions (which some other ISAs refer to as "unconditional branches"). RISC-V has two forms of jump instructions:

JAL
Jump and link. Jumps to PC + IMM (like branches, so this target is $br_tgt_pc, already assigned).
JALR
Jump and link register. Jumps to SRC1 + IMM.
"And link" refers to the fact that these instructions capture their original PC + 4 in a destination register, as you already coded in the ALU. (The link register is particularly useful for jumps that are used to implement function calls, which must return to the link address after function execution.)

Check the box for each completed step, to ensure none is skipped.

Compute $jalr_tgt_pc[31:0] (SRC1 + IMM).

Update the PC logic to select the correct $next_pc for JAL ($br_tgt_pc) and JALR ($jalr_tgt_pc). In the test program, JAL and JALR instructions should jump to the next subsequent instruction (as if not jumping at all), with the exception of the final JAL, which should jump to itself. Assuming x30 is also properly set to 1, this final JAL will result in the test reporting "Passed" in LOG and VIZ (though loads and stores are not working yet). Verify this behavior in VIZ.